Refactor: MenuItem 배열 분리 향후 수정할 가능성 있는 MyPage의 MenuList 관련 배열(MenuItem)을 따로 constant 디렉토리의 MenuItem.jsx 파일에서 관리
