+++ bovo_client/src/components/myPageListMenu/MenuList.jsx
@@ -1,8 +1,4 @@
 import PropTypes from "prop-types";
-import AccountBoxIcon from "@mui/icons-material/AccountBox";
-import EventIcon from "@mui/icons-material/Event";
-import InfoIcon from "@mui/icons-material/Info";
-import ExitToAppIcon from "@mui/icons-material/ExitToApp";
 import ChevronRightIcon from "@mui/icons-material/ChevronRight";
 import Box from '@mui/material/Box';
 import List from '@mui/material/List';
@@ -11,13 +7,7 @@ import ListItemIcon from '@mui/material/ListItemIcon';
 import ListItemText from '@mui/material/ListItemText';
 import styles from "./MenuList.module.css";
 import { Link } from "react-router-dom";
-
-const menuItem = [
-    { type: "link", text: "내 프로필", icon: <AccountBoxIcon sx={{fontSize: "3rem"}} />, path: "/mypage/myprofile" },
-    { type: "link", text: "독서 캘린더", icon: <EventIcon sx={{fontSize: "3rem"}} />, path: "/calendar" },
-    { type: "link", text: "서비스 정보", icon: <InfoIcon sx={{fontSize: "3rem"}} />, path: "/mypage/service-info" },
-    { type: "action", text: "로그 아웃", icon: <ExitToAppIcon sx={{fontSize: "3rem"}} />, action: "logout" },
-];
+import menuItem from "../../constant/MenuItem";
 
 const MenuList = ({ onLogout }) => {
     return (
